Move Flight Authorization helper out of password util
Keep util.py unchanged so the pre-existing PBKDF2 hash stays out of
this diff. Flight header parsing now lives in flight_headers.py.
